Move BaseComponent to package
