+++ app/components/__init__.py
@@ -1,4 +1,3 @@
-from .base import BaseComponent
 from .links import link_to, link_to_edit
 from .icons import icon
 from .action_badges import action_badge
@@ -6,7 +5,7 @@ from .dropzone import dropzone
 from .tables import table, search_box
 from .images import recipe_gallery_image
 
-__all__ = ["icon", "BaseComponent"]
+__all__ = ["icon"]
 
 
 def register_all_components(application):
