+++ app/packages/template_components/__init__.py
@@ -1,0 +1,3 @@
+from .core import BaseComponent
+
+__all__ = ["BaseComponent"]
